ModelParser: improve parsing of date fields
Rather than first matching the string against a regex
and then parsing it wiht ZonedDateTime.parse(),
just parse it, and fallback to StringFieldModel
if that does not work.

This is more flexible.
For example, perfectly valid dates with millis
like '2022-09-07T14:26:44.143Z', were previously
being converted to StringFieldModel,
because the regex was not considering millis.